requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-dbbackup==4.1.0
django-import-export==4.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.4
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
tzlocal==5.2
